+++ dpdk/examples/l3fwd-power/main.h
@@ -9,8 +9,8 @@
 #define MAX_LCORE_PARAMS 1024
 struct lcore_params {
 	uint16_t port_id;
-	uint8_t queue_id;
-	uint8_t lcore_id;
+	uint16_t queue_id;
+	uint32_t lcore_id;
 } __rte_cache_aligned;
 
 extern struct lcore_params *lcore_params;
